refactor: 뉴스기사 출처목록 병합 (#286)
* refactor: 출처목록 기능 병합

* refactor: 출처목록 테스트 병합

* chore: 출처목록 테스트 주석추가
