+++ src/main/java/com/team3/monew/controller/api/ArticleApi.java
@@ -18,11 +18,12 @@ import java.util.UUID;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@Tag(name = "뉴스 기사 관리", description = "뉴스 기사 관련 API")
+@Tag(name = "뉴스 기사 관리", description = "뉴스 기사 관리 API")
 public interface ArticleApi {
 
   String REQUEST_USER_ID_HEADER = "Monew-Request-User-ID";
@@ -45,7 +46,7 @@ public interface ArticleApi {
       @ApiResponse(responseCode = "200", description = "조회 성공"),
       @ApiResponse(responseCode = "400", description = "잘못된 요청 (헤더 누락, articleId 형식 오류, 삭제된 기사 조회 등)",
           content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-      @ApiResponse(responseCode = "404", description = "해당 ID에 맞는 뉴스기사가 존재하지 않음",
+      @ApiResponse(responseCode = "404", description = "해당 ID에 맞는 뉴스 기사가 존재하지 않음",
           content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
       @ApiResponse(responseCode = "500", description = "서버 내부 오류",
           content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
@@ -54,6 +55,15 @@ public interface ArticleApi {
       @RequestHeader(REQUEST_USER_ID_HEADER) UUID requestUserId,
       @PathVariable UUID articleId
   );
+
+  @Operation(summary = "뉴스 기사 출처 목록 조회", description = "뉴스 기사 출처 목록을 조회합니다.")
+  @ApiResponses({
+      @ApiResponse(responseCode = "200", description = "조회 성공"),
+      @ApiResponse(responseCode = "500", description = "서버 내부 오류",
+          content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+  })
+  @GetMapping("/sources")
+  ResponseEntity<List<String>> getArticleSources();
 
   @Operation(summary = "뉴스 기사 논리 삭제", description = "뉴스 기사를 논리적으로 삭제합니다.")
   @ApiResponses({
@@ -75,7 +85,7 @@ public interface ArticleApi {
   })
   ResponseEntity<Void> hardDeleteArticle(@PathVariable UUID articleId);
 
-  @Operation(summary = "뉴스 복구", description = "유실된 뉴스 기사를 복구.")
+  @Operation(summary = "뉴스 복구", description = "요청 기간의 뉴스 기사를 복구합니다.")
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "복구 성공"),
       @ApiResponse(responseCode = "400", description = "잘못된 요청(from, to 형식 오류)",
