+++ src/test/java/com/team3/monew/service/ArticleServiceTest.java
@@ -47,6 +47,7 @@ import com.team3.monew.repository.CommentRepository;
 import com.team3.monew.repository.NewsArticleRepository;
 import com.team3.monew.repository.NewsArticleRepository.ArticleCountInfo;
 import com.team3.monew.repository.NewsArticleRepository.ArticleLinkAndPublishedAt;
+import com.team3.monew.repository.NewsSourceRepository;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -100,6 +101,8 @@ class ArticleServiceTest {
   private ArticleInterestRepository articleInterestRepository;
   @Mock
   private CommentRepository commentRepository;
+  @Mock
+  private NewsSourceRepository newsSourceRepository;
   @Mock
   private ArticleViewService articleViewService;
   @Mock
@@ -814,6 +817,44 @@ class ArticleServiceTest {
       public Instant getPublishedAt() {
         return publishedAt;
       }
+    }
+  }
+
+  @Nested
+  @DisplayName("뉴스 기사 출처 목록을 조회한다")
+  class GetArticleSources {
+
+    @Test
+    @DisplayName("등록된 출처가 있으면 중복 없이 문자열 목록을 반환한다")
+    void shouldReturnDistinctArticleSources_whenNewsSourcesExist() {
+      // given
+      NewsSource naverPrimary = NewsSource.create("NAVER", NewsSourceType.NAVER,
+          "https://openapi.naver.com");
+      NewsSource naverSecondary = NewsSource.create("NAVER-SECONDARY", NewsSourceType.NAVER,
+          "https://openapi.naver.com/news");
+      NewsSource chosun = NewsSource.create("CHOSUN", NewsSourceType.CHOSUN,
+          "https://www.chosun.com");
+
+      given(newsSourceRepository.findAll()).willReturn(List.of(naverPrimary, naverSecondary, chosun));
+
+      // when
+      List<String> actual = articleService.getArticleSources();
+
+      // then
+      assertThat(actual).containsExactly("NAVER", "CHOSUN");
+    }
+
+    @Test
+    @DisplayName("등록된 출처가 없으면 빈 목록을 반환한다")
+    void shouldReturnEmptyList_whenNewsSourcesDoNotExist() {
+      // given
+      given(newsSourceRepository.findAll()).willReturn(List.of());
+
+      // when
+      List<String> actual = articleService.getArticleSources();
+
+      // then
+      assertThat(actual).isEmpty();
     }
   }
 }
